Remove concurrency from matching.Core. Move concurrency handling to matching.Engine
